requirements.txt
gradio
openai-whisper
torch
torchaudio
librosa
google-cloud-texttospeech
pydub
numpy
soundfile
nltk
audioread
ffmpeg-python
decorator
requests
openai
matplotlib
simplejson
scipy
hmmlearn
eyeD3
pydub
scikit_learn
tqdm
plotly
pandas
imblearn
pyaudioanalysis